requirements.txt
numpy==1.23.3
pandas==1.4.4
plotly==5.10.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
tenacity==8.0.1
